Cleanup pass: shared dedup, atomic social SQL, apiFetch helper, component splits
Six low-risk improvements bundled together; no UI/UX changes intended.

- Shared dedup: web `lib/shards.ts` re-exports shared shard tables; web-only
  presentation helpers stay. `BrowseCategory` type comes from shared (web
  keeps its own `CATEGORIES` array since labels/ordering intentionally
  differ from shared's `BROWSE_CATEGORIES`).
- Atomic like/bookmark: POST and DELETE now use raw
  `INSERT ... ON CONFLICT DO NOTHING` / `DELETE ... RETURNING id` instead of
  `prisma.$transaction`. Counter UPDATE only fires when a row actually
  changed. Eliminates the 500 on rapid double-click.
- View count off critical path: `registerBackgroundWork` defers the
  `UPDATE viewCount + 1` so the detail page responds without waiting on
  Neon.
- Ban session revoke: admin PATCH now deletes the banned user's sessions so
  the 60s Better Auth cookie cache can't keep them in.
- TanStack Router: drop `defaultPreloadStaleTime: 0`; per-query staleTime
  governs preloads now.
- React 19 ref-writes: `hotkeys.ts`, `drag-controller.tsx`, and
  `mod-search-grid.tsx` no longer write refs during render. Sync happens in
  an effect; mod-search-grid's uniqueName→index Map is `useMemo`d so it
  only rebuilds when `displayed` changes.
- `apiFetch` helper in `lib/api-client.ts` plus `ApiError` + shared
  `apiErrorMessage`. Migrated 11 call sites; raw-fetch stays for downloads
  that need headers/blob (`me-query.ts` export).
- Component splits: `item-sidebar.tsx` 1613 → 499 (split into stat-display,
  stats-panels, ability-icon, incarnon-controls, shard-controls,
  capacity-bar, lich-bonus-picker). `routes/create.tsx` 1010 → 683 (extracted
  EditorHeader, SearchPanel, useRivenDialog into `components/create-editor/`).
- Phase 6 cleanup: `SLUG_COLLISION_RETRIES` constant, deduped P2002 check
  via `hasPrismaCode`, chunked body parser (Uint8Array[] + single decode),
  GitHub OAuth missing-creds warn, `PlacedMod` flattener simplified,
  O(1) Map lookup in mod-search-grid keystroke handler.

Verified: `bun run build` (web) and `bunx tsc --noEmit` (api) both pass.
