+++ apps/api/src/lib/validate.ts
@@ -40,8 +40,7 @@ export async function parseJsonBody(
   }
 
   const reader = stream.getReader()
-  const decoder = new TextDecoder("utf-8", { fatal: true })
-  let raw = ""
+  const chunks: Uint8Array[] = []
   let total = 0
   try {
     while (true) {
@@ -53,9 +52,21 @@ export async function parseJsonBody(
         void reader.cancel()
         return { ok: false, response: jsonError("body_too_large", 413) }
       }
-      raw += decoder.decode(value, { stream: true })
+      chunks.push(value)
     }
-    raw += decoder.decode()
+  } catch {
+    return { ok: false, response: jsonError("invalid_body", 400) }
+  }
+
+  let raw: string
+  try {
+    const merged = new Uint8Array(total)
+    let offset = 0
+    for (const chunk of chunks) {
+      merged.set(chunk, offset)
+      offset += chunk.byteLength
+    }
+    raw = new TextDecoder("utf-8", { fatal: true }).decode(merged)
   } catch {
     return { ok: false, response: jsonError("invalid_body", 400) }
   }
