+++ apps/web/src/lib/build-actions.ts
@@ -1,6 +1,6 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query"
 
-import { API_URL } from "@/lib/constants"
+import { apiFetch, ApiError } from "@/lib/api-client"
 
 type ForkResponse = { id: string; slug: string }
 
@@ -8,13 +8,18 @@ export function useDeleteBuild(slug: string) {
   const qc = useQueryClient()
   return useMutation({
     mutationFn: async (): Promise<void> => {
-      const r = await fetch(`${API_URL}/builds/${encodeURIComponent(slug)}`, {
-        method: "DELETE",
-        credentials: "include",
-      })
-      if (r.status === 401) throw new Error("unauthorized")
-      if (r.status === 403) throw new Error("forbidden")
-      if (!r.ok && r.status !== 204) throw new Error("failed_delete")
+      try {
+        await apiFetch<void>(`/builds/${encodeURIComponent(slug)}`, {
+          method: "DELETE",
+        })
+      } catch (err) {
+        if (err instanceof ApiError) {
+          if (err.status === 401) throw new Error("unauthorized")
+          if (err.status === 403) throw new Error("forbidden")
+          throw new Error("failed_delete")
+        }
+        throw err
+      }
     },
     onSuccess: () => {
       qc.removeQueries({ queryKey: ["build", slug] })
@@ -27,14 +32,19 @@ export function useForkBuild(slug: string) {
   const qc = useQueryClient()
   return useMutation({
     mutationFn: async (): Promise<ForkResponse> => {
-      const r = await fetch(
-        `${API_URL}/builds/${encodeURIComponent(slug)}/fork`,
-        { method: "POST", credentials: "include" },
-      )
-      if (r.status === 401) throw new Error("unauthorized")
-      if (r.status === 404) throw new Error("not_found")
-      if (!r.ok) throw new Error("failed_fork")
-      return r.json()
+      try {
+        return await apiFetch<ForkResponse>(
+          `/builds/${encodeURIComponent(slug)}/fork`,
+          { method: "POST" },
+        )
+      } catch (err) {
+        if (err instanceof ApiError) {
+          if (err.status === 401) throw new Error("unauthorized")
+          if (err.status === 404) throw new Error("not_found")
+          throw new Error("failed_fork")
+        }
+        throw err
+      }
     },
     onSuccess: () => {
       qc.invalidateQueries({ queryKey: ["builds"] })
