+++ apps/web/src/lib/partner-builds-query.ts
@@ -5,7 +5,7 @@ import {
   useQueryClient,
 } from "@tanstack/react-query"
 
-import { API_URL } from "@/lib/constants"
+import { apiFetch, ApiError } from "@/lib/api-client"
 
 import type { BuildListItem } from "./builds-list-query"
 
@@ -17,13 +17,14 @@ export const partnerBuildsQuery = (slug: string) =>
   queryOptions({
     queryKey: ["build", slug, "partners"] as const,
     queryFn: async (): Promise<PartnerBuild[]> => {
-      const r = await fetch(
-        `${API_URL}/builds/${encodeURIComponent(slug)}/partners`,
-        { credentials: "include" },
-      )
-      if (!r.ok) throw new Error("failed_load_partners")
-      const data = (await r.json()) as PartnersResponse
-      return data.builds
+      try {
+        const data = await apiFetch<PartnersResponse>(
+          `/builds/${encodeURIComponent(slug)}/partners`,
+        )
+        return data.builds
+      } catch {
+        throw new Error("failed_load_partners")
+      }
     },
   })
 
@@ -31,13 +32,14 @@ export function useBuildSearch(q: string) {
   return useQuery({
     queryKey: ["builds", "search", q] as const,
     queryFn: async (): Promise<PartnerBuild[]> => {
-      const r = await fetch(
-        `${API_URL}/builds/search?q=${encodeURIComponent(q)}`,
-        { credentials: "include" },
-      )
-      if (!r.ok) throw new Error("failed_search")
-      const data = (await r.json()) as PartnersResponse
-      return data.builds
+      try {
+        const data = await apiFetch<PartnersResponse>(
+          `/builds/search?q=${encodeURIComponent(q)}`,
+        )
+        return data.builds
+      } catch {
+        throw new Error("failed_search")
+      }
     },
     enabled: q.trim().length >= 2,
     staleTime: 30_000,
@@ -48,13 +50,19 @@ export function useLinkPartner(ownerSlug: string) {
   const qc = useQueryClient()
   return useMutation({
     mutationFn: async (partner: PartnerBuild): Promise<void> => {
-      const r = await fetch(
-        `${API_URL}/builds/${encodeURIComponent(ownerSlug)}/partners/${encodeURIComponent(partner.slug)}`,
-        { method: "PUT", credentials: "include" },
-      )
-      if (r.status === 401) throw new Error("unauthorized")
-      if (r.status === 403) throw new Error("forbidden")
-      if (!r.ok && r.status !== 204) throw new Error("failed_link")
+      try {
+        await apiFetch<void>(
+          `/builds/${encodeURIComponent(ownerSlug)}/partners/${encodeURIComponent(partner.slug)}`,
+          { method: "PUT" },
+        )
+      } catch (err) {
+        if (err instanceof ApiError) {
+          if (err.status === 401) throw new Error("unauthorized")
+          if (err.status === 403) throw new Error("forbidden")
+          throw new Error("failed_link")
+        }
+        throw err
+      }
     },
     onMutate: async (partner) => {
       const key = ["build", ownerSlug, "partners"] as const
@@ -78,11 +86,14 @@ export function useUnlinkPartner(ownerSlug: string) {
   const qc = useQueryClient()
   return useMutation({
     mutationFn: async (partnerSlug: string): Promise<void> => {
-      const r = await fetch(
-        `${API_URL}/builds/${encodeURIComponent(ownerSlug)}/partners/${encodeURIComponent(partnerSlug)}`,
-        { method: "DELETE", credentials: "include" },
-      )
-      if (!r.ok && r.status !== 204) throw new Error("failed_unlink")
+      try {
+        await apiFetch<void>(
+          `/builds/${encodeURIComponent(ownerSlug)}/partners/${encodeURIComponent(partnerSlug)}`,
+          { method: "DELETE" },
+        )
+      } catch {
+        throw new Error("failed_unlink")
+      }
     },
     onMutate: async (partnerSlug) => {
       const key = ["build", ownerSlug, "partners"] as const
