+++ apps/web/src/lib/profile-query.ts
@@ -1,11 +1,11 @@
 import { queryOptions } from "@tanstack/react-query"
 import { notFound } from "@tanstack/react-router"
 
+import { apiFetch, ApiError } from "@/lib/api-client"
 import type {
   BuildListParams,
   BuildListResponse,
 } from "@/lib/builds-list-query"
-import { API_URL } from "@/lib/constants"
 
 export type ProfileBadges = {
   verified: boolean
@@ -37,13 +37,14 @@ export const profileQuery = (username: string) =>
   queryOptions({
     queryKey: ["profile", username.toLowerCase()],
     queryFn: async (): Promise<Profile> => {
-      const r = await fetch(
-        `${API_URL}/users/${encodeURIComponent(username)}`,
-        { credentials: "include" },
-      )
-      if (r.status === 404) throw notFound()
-      if (!r.ok) throw new Error("failed to load profile")
-      return r.json()
+      try {
+        return await apiFetch<Profile>(
+          `/users/${encodeURIComponent(username)}`,
+        )
+      } catch (err) {
+        if (err instanceof ApiError && err.status === 404) throw notFound()
+        throw new Error("failed to load profile")
+      }
     },
   })
 
@@ -57,11 +58,12 @@ export const profileBuildsQuery = (username: string, params: BuildListParams) =>
       if (params.q) q.set("q", params.q)
       if (params.category) q.set("category", params.category)
       const qs = q.toString() ? `?${q.toString()}` : ""
-      const r = await fetch(
-        `${API_URL}/users/${encodeURIComponent(username)}/builds${qs}`,
-        { credentials: "include" },
-      )
-      if (!r.ok) throw new Error("failed to load builds")
-      return r.json()
+      try {
+        return await apiFetch<BuildListResponse>(
+          `/users/${encodeURIComponent(username)}/builds${qs}`,
+        )
+      } catch {
+        throw new Error("failed to load builds")
+      }
     },
   })
