+++ apps/web/src/lib/shards.ts
@@ -1,32 +1,21 @@
-export type ShardColor =
-  | "crimson"
-  | "amber"
-  | "azure"
-  | "topaz"
-  | "violet"
-  | "emerald"
+export {
+  SHARD_COLORS,
+  SHARD_STATS,
+  getStatIndex,
+  getStatByIndex,
+} from "@arsenyx/shared/warframe/shards"
+export type {
+  ShardColor,
+  ShardStat,
+  PlacedShard,
+} from "@arsenyx/shared/warframe/types"
 
-export interface ShardStat {
-  name: string
-  baseValue: number
-  tauforgedValue: number
-  unit: string
-}
-
-export interface PlacedShard {
-  color: ShardColor
-  stat: string
-  tauforged: boolean
-}
-
-export const SHARD_COLORS: ShardColor[] = [
-  "crimson",
-  "amber",
-  "azure",
-  "topaz",
-  "violet",
-  "emerald",
-]
+import {
+  SHARD_STATS,
+  type ShardColor,
+  type ShardStat,
+  type PlacedShard,
+} from "@arsenyx/shared/warframe"
 
 export const SHARD_COLOR_NAMES: Record<ShardColor, string> = {
   crimson: "Crimson",
@@ -44,123 +33,6 @@ export const SHARD_CSS_COLORS: Record<ShardColor, string> = {
   topaz: "#ea580c",
   violet: "#7c3aed",
   emerald: "#059669",
-}
-
-export const SHARD_STATS: Record<ShardColor, ShardStat[]> = {
-  crimson: [
-    {
-      name: "Melee Critical Damage",
-      baseValue: 25,
-      tauforgedValue: 37.5,
-      unit: "%",
-    },
-    {
-      name: "Primary Status Chance",
-      baseValue: 25,
-      tauforgedValue: 37.5,
-      unit: "%",
-    },
-    {
-      name: "Secondary Critical Chance",
-      baseValue: 25,
-      tauforgedValue: 37.5,
-      unit: "%",
-    },
-    { name: "Ability Strength", baseValue: 10, tauforgedValue: 15, unit: "%" },
-    { name: "Ability Duration", baseValue: 10, tauforgedValue: 15, unit: "%" },
-  ],
-  amber: [
-    { name: "Initial Energy", baseValue: 30, tauforgedValue: 45, unit: "%" },
-    {
-      name: "Health Orb Effectiveness",
-      baseValue: 100,
-      tauforgedValue: 150,
-      unit: "%",
-    },
-    {
-      name: "Energy Orb Effectiveness",
-      baseValue: 50,
-      tauforgedValue: 75,
-      unit: "%",
-    },
-    { name: "Casting Speed", baseValue: 25, tauforgedValue: 37.5, unit: "%" },
-    {
-      name: "Parkour Velocity",
-      baseValue: 15,
-      tauforgedValue: 22.5,
-      unit: "%",
-    },
-  ],
-  azure: [
-    { name: "Health", baseValue: 150, tauforgedValue: 225, unit: "" },
-    { name: "Shield Capacity", baseValue: 150, tauforgedValue: 225, unit: "" },
-    { name: "Energy Max", baseValue: 50, tauforgedValue: 75, unit: "" },
-    { name: "Armor", baseValue: 150, tauforgedValue: 225, unit: "" },
-    { name: "Health Regen", baseValue: 5, tauforgedValue: 7.5, unit: "/s" },
-  ],
-  topaz: [
-    {
-      name: "Blast Kill Health",
-      baseValue: 1,
-      tauforgedValue: 2,
-      unit: " per kill",
-    },
-    {
-      name: "Blast Kill Shields",
-      baseValue: 5,
-      tauforgedValue: 7.5,
-      unit: " per kill",
-    },
-    {
-      name: "Heat Kill Crit Chance",
-      baseValue: 1,
-      tauforgedValue: 1.5,
-      unit: "% per kill",
-    },
-    {
-      name: "Radiation Ability Damage",
-      baseValue: 10,
-      tauforgedValue: 15,
-      unit: "%",
-    },
-  ],
-  violet: [
-    {
-      name: "Electricity Ability Damage",
-      baseValue: 10,
-      tauforgedValue: 15,
-      unit: "%",
-    },
-    {
-      name: "Primary Electricity Damage",
-      baseValue: 30,
-      tauforgedValue: 45,
-      unit: "%",
-    },
-    {
-      name: "Melee Critical Damage (Energy)",
-      baseValue: 25,
-      tauforgedValue: 37.5,
-      unit: "%",
-    },
-    { name: "Orb Conversion", baseValue: 20, tauforgedValue: 30, unit: "%" },
-  ],
-  emerald: [
-    {
-      name: "Toxin Status Damage",
-      baseValue: 30,
-      tauforgedValue: 45,
-      unit: "%",
-    },
-    { name: "Toxin Heal on Hit", baseValue: 2, tauforgedValue: 3, unit: " HP" },
-    {
-      name: "Corrosion Ability Damage",
-      baseValue: 10,
-      tauforgedValue: 15,
-      unit: "%",
-    },
-    { name: "Corrosion Max Stacks", baseValue: 2, tauforgedValue: 3, unit: "" },
-  ],
 }
 
 const SHARD_IMAGE_BASE = "https://wiki.warframe.com/images/thumb"
